Create more-info dialogs tests
