+++ .hass/config/.storage/core.restore_state
@@ -23,7 +23,7 @@
         }
       },
       "extra_data": null,
-      "last_seen": "2023-11-19T00:53:46.647507+00:00"
+      "last_seen": "2023-11-19T02:08:46.736835+00:00"
     },
     {
       "state": {
@@ -41,7 +41,7 @@
         }
       },
       "extra_data": null,
-      "last_seen": "2023-11-19T00:53:46.647507+00:00"
+      "last_seen": "2023-11-19T02:08:46.736835+00:00"
     }
   ]
 }
