+++ cypress/support/commands.ts
@@ -36,133 +36,14 @@
 //   }
 // }
 
-const NAME = 'Test';
-const PASSWORD = 'Test123!';
-
-Cypress.Commands.add('onboard', () => {
+Cypress.Commands.add('ingress', () => {
 
     cy.session('home-assistant', () => {
 
         cy
             .visit('http://localhost:8123')
             .location('pathname')
-            .should('eq', '/onboarding.html');
-
-            cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-welcome')
-            .shadow()
-            .find('ha-button')
-            .click();
-
-        cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-create-user')
-            .shadow()
-            .as('create-user');
-
-        cy
-            .get('@create-user')
-            .find('ha-form')
-            .shadow()
-            .as('form');
-
-        cy
-            .get('@form')
-            .find('ha-selector:nth-of-type(1)')
-            .shadow()
-            .find('ha-selector-text')
-            .shadow()
-            .find('ha-textfield')
-            .shadow()
-            .find('input[name="name"]')
-            .click()
-            .type(NAME);
-
-        cy
-            .get('@form')
-            .find('ha-selector:nth-of-type(3)')
-            .shadow()
-            .find('ha-selector-text')
-            .shadow()
-            .find('ha-textfield')
-            .shadow()
-            .find('input[name="password"]')
-            .click()
-            .type(PASSWORD);
-
-        cy
-            .get('@form')
-            .find('ha-selector:nth-of-type(4)')
-            .shadow()
-            .find('ha-selector-text')
-            .shadow()
-            .find('ha-textfield')
-            .shadow()
-            .find('input[name="password_confirm"]')
-            .click()
-            .type(PASSWORD);
-
-        cy
-            .get('@create-user')
-            .find('.footer mwc-button')
-            .click();
-
-        cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-core-config')
-            .shadow()
-            .find('onboarding-location')
-            .shadow()
-            .find('.footer mwc-button')
-            .click();
-
-        cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-core-config')
-            .shadow()
-            .as('form-country');
-
-        cy
-            .get('@form-country')
-            .find('ha-country-picker')
-            .shadow()
-            .find('ha-select')
-            .as('select');
-
-        cy
-            .get('@select')
-            .click();
-
-        cy
-            .get('@select')
-            .find('ha-list-item:nth-of-type(1)')
-            .click();
-
-        cy
-            .get('@form-country')
-            .find('.footer mwc-button')
-            .click();
-
-        cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-analytics')
-            .shadow()
-            .find('.footer mwc-button')
-            .click();
-
-        cy
-            .get('ha-onboarding')
-            .shadow()
-            .find('onboarding-integrations')
-            .shadow()
-            .find('.footer mwc-button')
-            .click();
+            .should('eq', '/lovelace/0');
 
     });
